% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseMSA.R, R/CrossedMSA.R,
%   R/NestedMSA.R
\name{anovaMSA}
\alias{anovaMSA}
\title{Generic for Anova MSA Study.}
\usage{
anovaMSA(object)

\S4method{anovaMSA}{BaseMSA}(object)

\S4method{anovaMSA}{CrossedMSA}(object)

\S4method{anovaMSA}{NestedMSA}(object)
}
\description{
Generic for Anova MSA Study.

Anova study for Base MSA

Anova study for Crossed MSA

Anova study for Nested MSA
}
