% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NestedMSA.R
\docType{methods}
\name{anovaStudy}
\alias{anovaStudy}
\alias{anovaStudy}
\title{Generic for Anova Study.}
\usage{
anovaStudy(object)

\S4method{anovaStudy}{NestedMSA}(object)
}
\description{
Generic for Anova Study.

Anova study for Nested MSA
}
