% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseMSA.R, R/CrossedMSA.R,
%   R/NestedMSA.R
\name{ggplotComponentOfVariationChart}
\alias{ggplotComponentOfVariationChart}
\title{Generic for plotting Component of Variation Chart with ggplot2}
\usage{
ggplotComponentOfVariationChart(object)

\S4method{ggplotComponentOfVariationChart}{BaseMSA}(object)

\S4method{ggplotComponentOfVariationChart}{CrossedMSA}(object)

\S4method{ggplotComponentOfVariationChart}{NestedMSA}(object)
}
\description{
Generic for plotting Component of Variation Chart with ggplot2

Components of Variation Chart with ggplot2

Components of Variation Chart with ggplot2

Components of Variation Chart with ggplot2
}
