% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CrossedMSA.R
\name{ggplotInteractionChart}
\alias{ggplotInteractionChart}
\title{Generic for plotting Interaction Chart with ggplot2}
\usage{
ggplotInteractionChart(object)

\S4method{ggplotInteractionChart}{CrossedMSA}(object)
}
\description{
Generic for plotting Interaction Chart with ggplot2

Interaction
}
