% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseMSA.R
\name{ggplotMeanChart}
\alias{ggplotMeanChart}
\title{Generic for plotting Mean Chart with grid ggplot2}
\usage{
ggplotMeanChart(object, ...)

\S4method{ggplotMeanChart}{BaseMSA}(object, gridLayout = FALSE, ...)
}
\description{
Generic for plotting Mean Chart with grid ggplot2

Mean Chart with ggplot2
}
