% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseMSA.R, R/CrossedMSA.R,
%   R/NestedMSA.R
\name{ggplotVariableByAppraiserChart}
\alias{ggplotVariableByAppraiserChart}
\title{Generic for plotting Variable by Appraiser Chart with ggplot2}
\usage{
ggplotVariableByAppraiserChart(object)

\S4method{ggplotVariableByAppraiserChart}{BaseMSA}(object)

\S4method{ggplotVariableByAppraiserChart}{CrossedMSA}(object)

\S4method{ggplotVariableByAppraiserChart}{NestedMSA}(object)
}
\description{
Generic for plotting Variable by Appraiser Chart with ggplot2

Variable by Appraiser with ggplot2

Variable by Appraiser with ggplot2

Variable by Appraiser with ggplot2
}
