% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseMSA.R, R/CrossedMSA.R,
%   R/NestedMSA.R
\name{ggplotVariableByPartChart}
\alias{ggplotVariableByPartChart}
\title{Generic for plotting Variable by Part Chart with ggplot2}
\usage{
ggplotVariableByPartChart(object)

\S4method{ggplotVariableByPartChart}{BaseMSA}(object)

\S4method{ggplotVariableByPartChart}{CrossedMSA}(object)

\S4method{ggplotVariableByPartChart}{NestedMSA}(object)
}
\description{
Generic for plotting Variable by Part Chart with ggplot2

Variable by Part

Variable by Part

Variable by Part
}
