% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseMSA.R
\name{initializeBaseMSA}
\alias{initializeBaseMSA}
\title{constructor for BaseMSA}
\usage{
\S4method{initialize}{BaseMSA}(
  .Object,
  ...,
  id,
  name,
  description,
  pro,
  characteristic,
  data,
  tolerance,
  sigma,
  alphaLim
)
}
\description{
This is the constructor.
}
