% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossedMSA.R
\name{initializeCrossedMSA}
\alias{initializeCrossedMSA}
\title{constructor for CrossedMSA}
\usage{
\S4method{initialize}{CrossedMSA}(
  .Object,
  ...,
  id,
  name,
  description,
  pro,
  characteristic,
  data,
  tolerance,
  sigma,
  alphaLim
)
}
\description{
This is the constructor.
}
