% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA.R
\docType{methods}
\name{initializeMSA}
\alias{initializeMSA}
\title{constructor for MSA}
\usage{
\S4method{initialize}{MSA}(.Object, ..., id, name, description, pro, variable,
  part, appraiser, data, usl = NA_real_, lsl = NA_real_,
  tolerance = NA_real_, sigma = 6, alphaLim = 0.05,
  errorTerm = "interaction", digits = 4)
}
\description{
This is the constructor.
}
