% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedMSA.R
\name{initializeNestedMSA}
\alias{initializeNestedMSA}
\title{constructor for NestedMSA}
\usage{
\S4method{initialize}{NestedMSA}(
  .Object,
  ...,
  id,
  name,
  description,
  pro,
  characteristic,
  data,
  tolerance,
  sigma,
  alphaLim
)
}
\description{
This is the constructor.
}
