% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NestedMSA.R
\name{plot}
\alias{plot}
\title{Generic for plotting MSA Study resume}
\usage{
plot(object)

\S4method{plot}{NestedMSA}(object)
}
\description{
Generic for plotting MSA Study resume

MSA resume Chart
}
