% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NestedMSA.R
\name{plotComponentOfVariation}
\alias{plotComponentOfVariation}
\title{Generic for plotting Component of Variation Chart}
\usage{
plotComponentOfVariation(object)

\S4method{plotComponentOfVariation}{NestedMSA}(object)

\S4method{plotControl}{NestedMSA}(object)
}
\description{
Generic for plotting Component of Variation Chart

Components of Variation Chart

Control Chart
}
