% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseMSA.R, R/CrossedMSA.R,
%   R/NestedMSA.R
\name{plotComponentOfVariationChart}
\alias{plotComponentOfVariationChart}
\title{Generic for plotting Component of Variation Chart}
\usage{
plotComponentOfVariationChart(object)

\S4method{plotComponentOfVariationChart}{BaseMSA}(object)

\S4method{plotComponentOfVariationChart}{CrossedMSA}(object)

\S4method{plotComponentOfVariationChart}{NestedMSA}(object)
}
\description{
Generic for plotting Component of Variation Chart

Components of Variation Chart

Components of Variation Chart

Components of Variation Chart
}
