% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CrossedMSA.R
\name{plotInteractionChart}
\alias{plotInteractionChart}
\title{Generic for plotting Interaction Chart}
\usage{
plotInteractionChart(object)

\S4method{plotInteractionChart}{CrossedMSA}(object)
}
\description{
Generic for plotting Interaction Chart

Interaction
}
