% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NestedMSA.R
\name{plotMean}
\alias{plotMean}
\title{Generic for plotting Mean Chart}
\usage{
plotMean(object)

\S4method{plotMean}{NestedMSA}(object)
}
\description{
Generic for plotting Mean Chart

Mean Chart
}
