% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NestedMSA.R
\name{plotRange}
\alias{plotRange}
\title{Generic for plotting Range Chart}
\usage{
plotRange(object)

\S4method{plotRange}{NestedMSA}(object)
}
\description{
Generic for plotting Range Chart

Range Chart
}
