% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseMSA.R, R/CrossedMSA.R,
%   R/NestedMSA.R
\name{plotRangeChart}
\alias{plotRangeChart}
\title{Generic for plotting Range Chart}
\usage{
plotRangeChart(object)

\S4method{plotRangeChart}{BaseMSA}(object)

\S4method{plotRangeChart}{CrossedMSA}(object)

\S4method{plotRangeChart}{NestedMSA}(object)
}
\description{
Generic for plotting Range Chart

Range Chart

Range Chart

Range Chart
}
