% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NestedMSA.R
\name{plotRar}
\alias{plotRar}
\title{Generic for plotting Component of Variation Chart}
\usage{
plotRar(object)

\S4method{plotRar}{NestedMSA}(object)
}
\description{
Generic for plotting Component of Variation Chart

Plot rar resume
}
