% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NestedMSA.R
\name{plotVariableByAppraiser}
\alias{plotVariableByAppraiser}
\title{Generic for plotting Variable by Appraiser Chart}
\usage{
plotVariableByAppraiser(object)

\S4method{plotVariableByAppraiser}{NestedMSA}(object)
}
\description{
Generic for plotting Variable by Appraiser Chart

Variable by Appraiser
}
