% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NestedMSA.R
\name{plotVariableByPart}
\alias{plotVariableByPart}
\title{Generic for plotting Variable by Part Chart}
\usage{
plotVariableByPart(object)

\S4method{plotVariableByPart}{NestedMSA}(object)
}
\description{
Generic for plotting Variable by Part Chart

Variable by Part
}
