% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseMSA.R, R/CrossedMSA.R,
%   R/NestedMSA.R
\name{plotVariableByPartChart}
\alias{plotVariableByPartChart}
\title{Generic for plotting Variable by Part Chart}
\usage{
plotVariableByPartChart(object)

\S4method{plotVariableByPartChart}{BaseMSA}(object)

\S4method{plotVariableByPartChart}{CrossedMSA}(object)

\S4method{plotVariableByPartChart}{NestedMSA}(object)
}
\description{
Generic for plotting Variable by Part Chart

Variable by Part

Variable by Part

Variable by Part
}
