% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseMSA.R, R/CrossedMSA.R,
%   R/NestedMSA.R
\name{rar}
\alias{rar}
\title{Generic for Gage rar MSA Study.}
\usage{
rar(object)

\S4method{rar}{BaseMSA}(object)

\S4method{rar}{CrossedMSA}(object)

\S4method{rar}{NestedMSA}(object)
}
\description{
Generic for Gage rar MSA Study.

Gage rar for Base MSA

Gage rar for Crossed MSA

Gage rar for Nested MSA
}
