% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISSQ.MSA.R
\docType{package}
\name{RISSQ.MSA}
\alias{RISSQ.MSA}
\title{Measurement System Analysis - R Infraestucture for Statistical Standards on Quality}
\description{
Measurement System Analysis - R Infraestucture for Statistical Standards on Quality
}
