% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossedMSA.R, R/MSA.R
\name{summary}
\alias{summary}
\title{Summarize a MSA object}
\usage{
\S4method{summary}{CrossedMSA}(object, ...)

\S4method{summary}{MSA}(object, ...)
}
\description{
Summarize a MSA object

Summarize a MSA object
}
