% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NestedMSA.R
\docType{methods}
\name{anovaMSA}
\alias{anovaMSA}
\alias{anovaMSA}
\title{Generic for Anova MSA Study.}
\usage{
anovaMSA(object)

\S4method{anovaMSA}{NestedMSA}(object)
}
\description{
Generic for Anova MSA Study.

Anova study for Nested MSA
}
