% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ggplotMeanChart}
\alias{ggplotMeanChart}
\title{Generic for plotting Mean Chart with grid ggplot2}
\usage{
ggplotMeanChart(object, ...)
}
\description{
Generic for plotting Mean Chart with grid ggplot2
}
