% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseMSA.R, R/CrossedMSA.R,
%   R/NestedMSA.R
\name{ggplotRangeChart}
\alias{ggplotRangeChart}
\title{Generic for plotting Range Chart with ggplot2}
\usage{
ggplotRangeChart(object, ...)

\S4method{ggplotRangeChart}{BaseMSA}(object, gridLayout = FALSE, ...)

\S4method{ggplotRangeChart}{CrossedMSA}(object, gridLayout = FALSE, ...)

\S4method{ggplotRangeChart}{NestedMSA}(object, gridLayout = FALSE, ...)
}
\description{
Generic for plotting Range Chart with ggplot2

Range Chart with ggplot2

Range Chart with ggplot2

Range Chart with ggplot2
}
