% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA.R
\name{initializeMSA}
\alias{initializeMSA}
\title{constructor for MSA}
\usage{
\S4method{initialize}{MSA}(
  .Object,
  ...,
  id,
  name,
  description,
  pro,
  characteristic,
  data,
  tolerance,
  sigma,
  alphaLim
)
}
\description{
This is the constructor.
}
