% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA.R
\docType{methods}
\name{initializeMSA}
\alias{initializeMSA}
\title{constructor for MSA}
\usage{
\S4method{initialize}{MSA}(.Object, ..., id, name, description, pro, variable,
  part, appraiser, data, usl, lsl, tolerance, sigma, alphaLim, digits)
}
\description{
This is the constructor.
}
