% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedMSA.R
\docType{methods}
\name{initializeNestedMSA}
\alias{initializeNestedMSA}
\title{constructor for NestedMSA}
\usage{
\S4method{initialize}{NestedMSA}(.Object, ..., id, name, description, pro,
  variable, part, appraiser, data, usl, lsl, tolerance, sigma, alphaLim, digits)
}
\description{
This is the constructor.
}
