% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedMSA.R
\docType{methods}
\name{initializeNestedMSA}
\alias{initializeNestedMSA}
\title{constructor for NestedMSA}
\usage{
\S4method{initialize}{NestedMSA}(.Object, ..., id, name, description,
  pro = NA, variable, part, appraiser, data, usl = NA_real_,
  lsl = NA_real_, tolerance = NA_real_, sigma = 6, alphaLim = 0.05,
  errorTerm = "interaction", digits = 4)
}
\description{
This is the constructor.
}
