% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseMSA.R, R/CrossedMSA.R, R/NestedMSA.R
\name{plotGridMeanChart}
\alias{plotGridMeanChart}
\title{Mean Chart with ggplot2}
\usage{
\S4method{ggplotMeanChart}{BaseMSA}(object, gridLayout = FALSE, ...)

\S4method{ggplotMeanChart}{CrossedMSA}(object, gridLayout = FALSE, ...)

\S4method{ggplotMeanChart}{NestedMSA}(object, gridLayout = FALSE, ...)
}
\description{
Mean Chart with ggplot2

Mean Chart with ggplot2

Mean Chart with ggplot2
}
