% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseMSA.R, R/CrossedMSA.R,
%   R/NestedMSA.R
\name{plotMeanChart}
\alias{plotMeanChart}
\title{Generic for plotting Mean Chart}
\usage{
plotMeanChart(object)

\S4method{plotMeanChart}{BaseMSA}(object)

\S4method{plotMeanChart}{CrossedMSA}(object)

\S4method{plotMeanChart}{NestedMSA}(object)
}
\description{
Generic for plotting Mean Chart

Mean Chart

Mean Chart

Mean Chart
}
