% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CrossedMSA.R, R/MSA.R
\name{plotRar}
\alias{plotRar}
\title{Generic for plotting Component of Variation Chart}
\usage{
plotRar(object)

\S4method{plotRar}{CrossedMSA}(object)

\S4method{plotRar}{MSA}(object)
}
\description{
Generic for plotting Component of Variation Chart

Plot rar resume

Plot rar resume
}
