% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseMSA.R, R/CrossedMSA.R,
%   R/NestedMSA.R
\name{plotVariableByAppraiserChart}
\alias{plotVariableByAppraiserChart}
\title{Generic for plotting Variable by Appraiser Chart}
\usage{
plotVariableByAppraiserChart(object)

\S4method{plotVariableByAppraiserChart}{BaseMSA}(object)

\S4method{plotVariableByAppraiserChart}{CrossedMSA}(object)

\S4method{plotVariableByAppraiserChart}{NestedMSA}(object)
}
\description{
Generic for plotting Variable by Appraiser Chart

Variable by Appraiser

Variable by Appraiser

Variable by Appraiser
}
