% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NestedMSA.R
\docType{methods}
\name{rar}
\alias{rar}
\alias{rar}
\title{Generic for Gage rar MSA Study.}
\usage{
rar(object)

\S4method{rar}{NestedMSA}(object)
}
\description{
Generic for Gage rar MSA Study.

Gage rar for Nested MSA
}
